# specs/003-migrate-supabase-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Content Hub Manager - SupaBase Auth + Assets API
  version: 0.1.0
paths:
  /auth/signup:
    post:
      summary: Create a new user (signup)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Created
  /auth/login:
    post:
      summary: Login (returns session token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT